Server/NFTGen/requirements.txt
Flask~=3.0.3
web3~=6.16.0
python-dotenv~=1.0.1